+++ app/src/main/java/com/example/nimble/catalog/CatalogFragment.kt
@@ -1,0 +1,21 @@
+package com.example.nimble.catalog
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.nimble.R
+
+
+class CatalogFragment : Fragment() {
+
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_catalog, container, false)
+    }
+
+}
