+++ app/src/main/res/layout/fragment_catalog.xml
@@ -3,8 +3,32 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CatalogFragment">
+    tools:context=".catalog.CatalogFragment"
+    android:background="@color/background">
+
+    <LinearLayout
+        android:id="@+id/nav_Search"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:orientation="vertical"
+        android:padding="10dp"
+        android:background="@color/actionbarcolor">
+
+        <androidx.appcompat.widget.SearchView
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            />
 
 
+    </LinearLayout>
+
+
+    <androidx.core.widget.NestedScrollView
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:fillViewport="true">
+
+
+    </androidx.core.widget.NestedScrollView>
 
 </FrameLayout>
