+++ app/src/main/res/layout/fragment_pannier.xml
@@ -3,7 +3,9 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PannierFragment">
+    tools:context=".PannierFragment"
+    android:background="@color/background"
+    >
